refactor(compiler): better facade pattern for file system
